Ensure Controller::$methods is updated on enable() and disable()
If Controller::$methods isn't updated correctly on CrudAction::enable/disbale
it can have some interesting byproducts.

E.g. if you call disable() on an action, you should be served with the normal
MissingMethodException as if the method really didn't exist in the controller
rather than a 'action not mapped' CakeException

If enable() doesn't add the action to the methods array, components like Auth
won't be able to check authorization for the action, since they depend on
the methods property to figure out what to protect.

This was implemented in v2, but was lost in the refactoring