Make mapResources test more flexible
Unfortunately, the number of controllers in the core's test_app folder
changes with different versions. Rather than create a copy of the
controller, or by messing around with the cache class to "mock"
controller objects - make the test able to cope with any number of
controllers. Even though this goes against the idea of always having
static expectations for robustness.