Initial pass at a BulkAction base handler
This will allow us to implement `_bulk` handlers for actions that occur from `index`-like pages. Any subclass that implements a `_bulk` method will be given a $query object that can be used to reference a set of database objects. The classes can then do whatever they like - such as publish all posts, or delete records.

This sort of method is more useful for admin modules than for apis, but is a common task in application development