Combine AddAction and EditAction logic into BaseCreateAction
This change provides a single interface into creating/updating
records. The logic from both classes was highly similar, with only
the retrieval of the initial GET subject and the entity being posted
being different.

One possibility this change introduces is the ability to have a
`CreateOrUpdateAction`, wherein the entity will be created regardless
of it's initial existence. Users can simply subclass the
`BaseCreateAction` class and define the two methods necessary to
handle the cases they wish to handle.