Make ControllerTrait::isAction() return a dummy closure for mapped actions.
This is to prevent the ControllerFactory from trying to type cast the method args.
It was just by dumb luck that this didn't cause problems earlier as we only have
actions with a single argument and 1st argument of CrudComponent::execute() is typed
as nullable string. Any custom action with non array 2nd argument would be broken.